Convert GeneratedDocumentPublisher to an interface
