+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/GeneratedDocumentSynchronizer.cs
@@ -10,12 +10,12 @@ namespace Microsoft.AspNetCore.Razor.LanguageServer;
 
 internal class GeneratedDocumentSynchronizer : DocumentProcessedListener
 {
-    private readonly GeneratedDocumentPublisher _publisher;
+    private readonly IGeneratedDocumentPublisher _publisher;
     private readonly IDocumentVersionCache _documentVersionCache;
     private readonly ProjectSnapshotManagerDispatcher _dispatcher;
 
     public GeneratedDocumentSynchronizer(
-        GeneratedDocumentPublisher publisher,
+        IGeneratedDocumentPublisher publisher,
         IDocumentVersionCache documentVersionCache,
         ProjectSnapshotManagerDispatcher dispatcher)
     {
