+++ src/Razor/test/Microsoft.AspNetCore.Razor.LanguageServer.Test/GeneratedDocumentSynchronizerTest.cs
@@ -60,22 +60,18 @@ public class GeneratedDocumentSynchronizerTest : LanguageServerTestBase
         Assert.True(_publisher.PublishedHtml);
     }
 
-    private class TestGeneratedDocumentPublisher : GeneratedDocumentPublisher
+    private class TestGeneratedDocumentPublisher : IGeneratedDocumentPublisher
     {
-        public override void Initialize(ProjectSnapshotManagerBase projectManager)
-        {
-        }
-
         public bool PublishedCSharp { get; private set; }
 
         public bool PublishedHtml { get; private set; }
 
-        public override void PublishCSharp(ProjectKey projectKey, string filePath, SourceText sourceText, int hostDocumentVersion)
+        public void PublishCSharp(ProjectKey projectKey, string filePath, SourceText sourceText, int hostDocumentVersion)
         {
             PublishedCSharp = true;
         }
 
-        public override void PublishHtml(ProjectKey projectKey, string filePath, SourceText sourceText, int hostDocumentVersion)
+        public void PublishHtml(ProjectKey projectKey, string filePath, SourceText sourceText, int hostDocumentVersion)
         {
             PublishedHtml = true;
         }
